+++ sites/funmooc/src/backend/funmooc/settings.py
@@ -252,6 +252,8 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
                 environ_name="EDX_COURSE_REGEX",
                 environ_prefix=None,
             ),
+            # Synchronization
+            "COURSE_RUN_SYNC_NO_UPDATE_FIELDS": ["languages"],
             "JS_COURSE_REGEX": values.Value(
                 r"^.*/courses/(?<course_id>.*)/info/?$",
                 environ_name="EDX_JS_COURSE_REGEX",