+++ sites/funcampus/src/backend/funcampus/settings.py
@@ -12,6 +12,8 @@ from configurations import Configuration, values
 from richie.apps.courses.settings.mixins import RichieCoursesConfigurationMixin
 from sentry_sdk.integrations.django import DjangoIntegration
 
+from base.utils import merge_dict
+
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 DATA_DIR = os.path.join("/", "data")
 
@@ -233,6 +235,15 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         }
     ]
 
+    # Placeholders
+    CMS_PLACEHOLDER_CONF_OVERRIDES = {
+        "courses/cms/course_detail.html course_teaser": {
+            "name": _("Teaser"),
+            "plugins": ["LTIConsumerPlugin"],
+            "limits": {"LTIConsumerPlugin": 1},
+        },
+    }
+
     MIDDLEWARE = (
         "richie.apps.core.cache.LimitBrowserCacheTTLHeaders",
         "cms.middleware.utils.ApphookReloadMiddleware",
@@ -264,6 +275,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         "richie.plugins.glimpse",
         "richie.plugins.html_sitemap",
         "richie.plugins.large_banner",
+        "richie.plugins.lti_consumer",
         "richie.plugins.nesteditem",
         "richie.plugins.plain_text",
         "richie.plugins.section",
@@ -449,6 +461,31 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     )
     RICHIE_ES_STATE_WEIGHTS = values.ListValue(None)
 
+    # LTI Content
+    RICHIE_LTI_PROVIDERS = {
+        "marsha": {
+            "oauth_consumer_key": values.Value(
+                "InsecureOauthConsumerKey",
+                environ_name="LTI_TEST_OAUTH_CONSUMER_KEY",
+                environ_prefix=None,
+            ),
+            "shared_secret": values.Value(
+                "InsecureSharedSecret",
+                environ_name="LTI_TEST_SHARED_SECRET",
+                environ_prefix=None,
+            ),
+            "base_url": values.Value(
+                "https://marsha\.education/lti/videos/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}",  # noqa
+                environ_name="LTI_TEST_BASE_URL",
+                environ_prefix=None,
+            ),
+            "display_name": "Marsha Video",
+            "is_base_url_regex": True,
+            "automatic_resizing": True,
+            "inline_ratio": 0.5625,
+        }
+    }
+
     # Cache
     CACHES = values.DictValue(
         {
@@ -558,6 +595,11 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
             with sentry_sdk.configure_scope() as scope:
                 scope.set_extra("application", "backend")
 
+        # Customize DjangoCMS placeholders configuration
+        cls.CMS_PLACEHOLDER_CONF = merge_dict(
+            cls.CMS_PLACEHOLDER_CONF, cls.CMS_PLACEHOLDER_CONF_OVERRIDES
+        )
+
 
 class Development(Base):
     """
@@ -610,7 +652,8 @@ class Development(Base):
 
 class Test(Base):
     """Test environment settings"""
-    STATICFILES_STORAGE = 'django.contrib.staticfiles.storage.StaticFilesStorage'
+
+    STATICFILES_STORAGE = "django.contrib.staticfiles.storage.StaticFilesStorage"
 
 
 class ContinuousIntegration(Test):