+++ sites/funmooc/CHANGELOG.md
@@ -13,6 +13,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add "Types" category as search filter
 - Use custom views to handle errors (400, 403, 404, 500)
 
+### Changed
+
+- Prevent course run languages update when synchronization hook is triggered
+
 ## [1.0.0] - 2021-02-05
 
 ### Added