+++ sites/funmooc/CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add "Types" category as search filter
 - Use custom views to handle errors (400, 403, 404, 500)
+- Enable LTIConsumerPlugin on `course_teaser` placeholder
 
 ### Changed